Enable loading component on bottles/entities lists
