+++ apps/web-next/src/app/(bottles-sidebar)/bottles/loading.tsx
@@ -1,0 +1,5 @@
+import Spinner from "@peated/web/components/spinner";
+
+export default function Loading() {
+  return <Spinner />;
+}
