+++ apps/web-next/src/app/(entities-sidebar)/entities/layout.tsx
@@ -1,0 +1,16 @@
+import Layout from "@peated/web/components/layout";
+import { type Metadata } from "next";
+import { type ReactNode } from "react";
+import EntityListSidebar from "./rightSidebar";
+
+export const metadata: Metadata = {
+  title: "Search Whisky Brands, Bottler, and Distillers",
+};
+
+export default async function BottlesLayout({
+  children,
+}: {
+  children: ReactNode;
+}) {
+  return <Layout rightSidebar={<EntityListSidebar />}>{children}</Layout>;
+}
