+++ apps/web-next/src/app/(entities-sidebar)/entities/rightSidebar.tsx
@@ -8,6 +8,7 @@ import { useSearchParams } from "next/navigation";
 
 export default function EntityListSidebar() {
   const searchParams = useSearchParams();
+
   return (
     <div className="mt-8 flex flex-col overflow-y-auto bg-slate-950 px-6 py-4">
       <ul role="list" className="flex flex-auto flex-col gap-y-7">
@@ -17,17 +18,15 @@ export default function EntityListSidebar() {
           </Button>
         </li>
         <FilterSidebarSection
-          searchParams={searchParams}
           name="type"
           options={ENTITY_TYPE_LIST.map((k) => [k, toTitleCase(k)])}
         />
         <FilterSidebarSection
-          searchParams={searchParams}
           name="country"
           options={MAJOR_COUNTRIES.map((k) => [k, k])}
         />
         {searchParams.get("region") ? (
-          <FilterSidebarSection searchParams={searchParams} name="region" />
+          <FilterSidebarSection name="region" />
         ) : null}
       </ul>
     </div>
