+++ apps/web-next/src/components/header.tsx
@@ -22,7 +22,7 @@ export default function Header({
       >
         <div
           className={classNames(
-            "fixed left-0 right-0 z-10",
+            "fixed left-0 right-0 z-50",
             color === "primary"
               ? "main-gradient backdrop-blur"
               : "border-b border-b-slate-700 bg-slate-950",
